+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 Change Log
 ==========
 
+Unreleased
+----------
+- Support for model constraints
+- Support for data skipping indexes
+
 v2.0.1
 ------
 - Remove unnecessary import of `six`